requirements.txt
Django==5.0.2
Pillow==10.2.0
python-dotenv==1.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
psycopg2-binary==2.9.9
django-filter==23.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
WeasyPrint==60.1 